added support for older versions of openssh
Older versions (like the version used in the affected debian jessie)
don't support "-E" yet and just output the MD5 hash of the key. So
basically the check can just be omitted because the output is compatible
to the existing script.

As a precaution to avoid false positives due to unknown ssh-keygen
versions with a completely different output format I added a test
for a random public key that I just generated instead. This way the
script can be fairly sure it is able to detec an offending key.